Debug.c remove all errored/warninged code from API change
This shouldn't break the functionality of the driver since it is debug
functionality, but nonetheless these call should be fixed so they
compile.

Specifically our problem children are create_proc_entry() and
create_read_proc_entry() which are now deprecated and removed from the
kernel header file.

Possible solutions:
1. Use the replacement functions instead (different arg list)
2. Copy these functions from an older kernel file (may cause weird
behavior)