Added troubleshooting section
There were fixes suggested on various forums to configure the driver, these fixes have mixed results.  One of the options common was debug=2. The debug option was removed in newer versions of the driver so when the new drivers are loaded it fails because of the unsupported option in the config.