verify_branch try to get a git checkout if one doesn't exist
If the user downloads the source code as a zip file then there is no git
repo.  This is a problem because they usually download the source on the
master branch.

If we're not in a git repo, we'll try to install the dependencies and
create a git checkout (with permission of course)